+++ src/api/get-nonce.js
@@ -21,7 +21,13 @@ export default function GetNonce(apiInstance) {
             .then((response) => {
                 const resData = response.data;
                 const nonce = apiInstance.defaults.apiType === API_TYPE_GATE ? resData.nonce : resData.transaction_count;
-                return Number(nonce) + 1;
+                const newNonce = Number(nonce) + 1;
+                if (Number.isNaN(newNonce)) {
+                    // eslint-disable-next-line unicorn/prefer-type-error
+                    throw new Error(`Can't get nonce, API returns '${nonce}'`);
+                }
+
+                return newNonce;
             });
     };
 }