+++ src/tx-data/pool-remove-liquidity.js
@@ -1,7 +1,6 @@
 import {TxDataRemoveLiquidity} from 'minterjs-tx';
 import {convertFromPip, convertToPip, toBuffer} from 'minterjs-util';
 import {proxyNestedTxData, bufferToInteger, integerToHexString, validateAmount, validateUint} from '../utils.js';
-// import {convertToPip} from 'minterjs-util/src/converter.js';
 
 /**
  * @param {number|string} coin0 - coin id