+++ src/tx-data/vote-update.js
@@ -1,22 +1,24 @@
-import {TxDataSetHaltBlock} from 'minterjs-tx';
+import {TxDataVoteUpdate} from 'minterjs-tx';
 import {toBuffer, publicToString} from 'minterjs-util';
 import {bufferToInteger, integerToHexString, proxyNestedTxData, validateUint, validatePublicKey} from '../utils.js';
 
 
 /**
- *
+ * @param {string} version
  * @param {string} publicKey
  * @param {number|string} height
  * @constructor
  */
-export default function SetHaltBlockTxData({publicKey, height}) {
+export default function VoteUpdateTxData({version, publicKey, height}) {
     validatePublicKey(publicKey, 'publicKey');
     validateUint(height, 'height');
 
+    this.version = version;
     this.publicKey = publicKey;
     this.height = height;
 
-    this.txData = new TxDataSetHaltBlock({
+    this.txData = new TxDataVoteUpdate({
+        version: Buffer.from(version.toString(), 'utf-8'),
         publicKey: toBuffer(publicKey),
         height: integerToHexString(height),
     });
@@ -25,12 +27,14 @@ export default function SetHaltBlockTxData({publicKey, height}) {
 }
 
 /**
+ * @param {Buffer|string} version
  * @param {Buffer|string} publicKey
  * @param {Buffer|string|number} height
- * @return {SetHaltBlockTxData}
+ * @return {VoteUpdateTxData}
  */
-SetHaltBlockTxData.fromBufferFields = function fromBufferFields({publicKey, height}) {
-    return new SetHaltBlockTxData({
+VoteUpdateTxData.fromBufferFields = function fromBufferFields({version, publicKey, height}) {
+    return new VoteUpdateTxData({
+        version: toBuffer(version).toString('utf-8'),
         publicKey: publicToString(publicKey),
         height: bufferToInteger(toBuffer(height)),
     });
@@ -38,8 +42,8 @@ SetHaltBlockTxData.fromBufferFields = function fromBufferFields({publicKey, heig
 
 /**
  * @param {Buffer|string} data
- * @return {SetHaltBlockTxData}
+ * @return {VoteUpdateTxData}
  */
-SetHaltBlockTxData.fromRlp = function fromRlp(data) {
-    return SetHaltBlockTxData.fromBufferFields(new TxDataSetHaltBlock(data));
+VoteUpdateTxData.fromRlp = function fromRlp(data) {
+    return VoteUpdateTxData.fromBufferFields(new TxDataVoteUpdate(data));
 };